==> notebooks_and_code/1_2a - Loading Mou et al datasets.ipynb <==
### 1. Loading and preprocessing training data from Mou et al.
### 2. Creating substrate representations
### 3. Creating enzyme representations

import pandas as pd
import numpy as np
import random
from os.path import join
import os
import re
import sys
import time
from rdkit import Chem
from rdkit import DataStructs
from rdkit.Chem import AllChem
from Bio import SeqIO
import warnings
import torch
warnings.filterwarnings("ignore")

sys.path.append('.\\additional_code')
from data_preprocessing import *

CURRENT_DIR = os.getcwd()
print(CURRENT_DIR)

### 1. Loading and preprocessing training data from Mou et al.

df = pd.read_excel(join(CURRENT_DIR, ".." ,"data", "Mou_data", 'nitrilase_descriptors.xlsx'),sheet_name = 'neu50').copy()
display(df.head())
df = pd.DataFrame(data= {"enzyme" : df["protein"], "metabolite" : df["nitriles"],
                        "activity" : df["activity"]})
df

### 2. Creating substrate representations

metabolites = list(set(df["metabolite"]))

met_dict = {'2,6-Dichlorobenzonitrile' : "2_6-dichlorobenzonitrile",
           '⍺-Methylbenzyl cyanide' : "alpha-methylbenzylcyanide",
           '4-(Dimethylamino)benzonitrile' :"4-Dimethylamino-benzonitrile"}

df["ECFP"] = np.nan

for met in metabolites:
    met2 = met
    try:
        met2 = met_dict[met]
    except KeyError:
        pass
    
    mol = Chem.MolFromMolFile(join(CURRENT_DIR, ".." ,"data", "Mou_data","nitrile_coordinates", met2 + ".mol"))
    ecfp = AllChem.GetMorganFingerprintAsBitVect(mol, 3, nBits=1024).ToBitString()
    
    df["ECFP"].loc[df["metabolite"] == met] = ecfp

df

### 3. Creating enzyme representations

enzymes = list(set(df["enzyme"]))

enzy_str = ""
for enz in enzymes:
    enzy_str += enz +","
enzy_str

https://www.rcsb.org/downloads/fasta

from Bio import SeqIO

df["Sequence"] = np.nan

pdb_files = os.listdir(join(CURRENT_DIR, ".." ,"data", "Mou_data", "pdb_files"))
for pdb in pdb_files:
    for i, record in enumerate(SeqIO.parse(join(CURRENT_DIR, ".." ,"data", "Mou_data", "pdb_files", pdb), "pdb-atom")):
        if i ==0:
            AA =str(record.seq)
            print(pdb, AA)
        df["Sequence"].loc[df["enzyme"] == pdb.split("_")[0]] = AA

Creating fasta file with all sequences:

ofile = open(join(CURRENT_DIR, ".." ,"data", "Mou_data", "all_sequences_Mou.fasta"), "w")
for ind in df.index:
    seq = df["Sequence"][ind]
    if not pd.isnull(seq):
        seq_end = seq.find("#")
        seq = seq[:seq_end]
        ofile.write(">" + str(ind) + "\n" + seq  + "\n")
ofile.close()

Mapping enzyme representations:

import torch
rep_dict = torch.load(join(CURRENT_DIR, ".." ,"data", "Mou_data", "all_sequences_Mou.pt"))


df["ESM1b_ts"] = ""
for ind in df.index:
    df["ESM1b_ts"][ind] = rep_dict[str(ind) +".pt"]

df.to_pickle(join(CURRENT_DIR, ".." ,"data", "Mou_data", "Mou_df.pkl"))
