==> notebooks_and_code/2_1 - Training gradient boosting models with additional data from Yang et al.ipynb <==
# Training gradient boosting model for enzyme-substrate pair prediction with ESM-1b-vectors

### 1. Loading and preprocessing data for model training and evaluation
### 2. Hyperparameter optimization using a 5-fold cross-validation (CV)
### 3. Training and validating the final model

import pandas as pd
import numpy as np
import random
import pickle
import sys
import os
import logging
from os.path import join
from sklearn.model_selection import KFold
#from hyperopt import fmin, tpe, hp, Trials, rand
import xgboost as xgb
from sklearn.metrics import roc_auc_score

sys.path.append('.\\additional_code')
#from data_preprocessing import *

CURRENT_DIR = os.getcwd()
print(CURRENT_DIR)

## 1. Loading and preprocessing data for model training and evaluation

### (a) Loading data:

df_test  = pd.read_pickle(join(CURRENT_DIR, ".." ,"data","splits", "df_test_with_ESM1b_ts.pkl"))
df_test = df_test.loc[df_test["ESM1b_ts"] != ""]
df_test.reset_index(inplace = True, drop = True)

df_Mou  = pd.read_pickle(join(CURRENT_DIR, ".." ,"data", "Yang_data", "Yang_df.pkl"))
df_Mou = df_Mou.loc[df_Mou["ESM1b_ts"] != ""]
df_Mou.reset_index(inplace = True, drop = True)

df_Berry  = pd.read_pickle(join(CURRENT_DIR, ".." ,"data", "Yang_data", "Yang_validation_Berry_df.pkl"))
df_Berry = df_Berry.loc[df_Berry["ESM1b_ts"] != ""]
df_Berry.reset_index(inplace = True, drop = True)

df_Oat  = pd.read_pickle(join(CURRENT_DIR, ".." ,"data", "Yang_data", "Yang_validation_Oat_df.pkl"))
df_Oat = df_Oat.loc[df_Oat["ESM1b_ts"] != ""]
df_Oat.reset_index(inplace = True, drop = True)

np.mean(df_Mou["Binding"]), np.mean(df_Berry["Binding"]), np.mean(df_Oat["Binding"])

Loading new dataset:

def create_input_and_output_data(df):
    X = ();
    y = ();
    
    for ind in df.index:
        emb = df["ESM1b_ts"][ind]
        ecfp = np.array(list(df["ECFP"][ind])).astype(int)
                
        X = X +(np.concatenate([ecfp, emb]), );
        y = y + (df["Binding"][ind], );

    return(np.array(X),np.array(y))

feature_names =  ["ECFP_" + str(i) for i in range(1024)]
feature_names = feature_names + ["ESM1b_ts_" + str(i) for i in range(1280)]

#import matplotlib.pyplot as plt

### Adding Yang et al. data to the training set:


df_train = pd.read_pickle(join(CURRENT_DIR, ".." ,"data",
                               "splits", "df_train_with_ESM1b_ts.pkl"))
df_train = df_train.loc[df_train["ESM1b_ts"] != ""]
df_train.reset_index(inplace = True, drop = True)


train_X, train_y =  create_input_and_output_data(df = df_train)
test_X, test_y =  create_input_and_output_data(df = df_test)

df_test_new = df_Mou.copy()
test_new_X, test_new_y =  create_input_and_output_data(df = df_test_new)



X_train_Mou = test_new_X
y_train_Mou = test_new_y

train_X = np.concatenate([train_X, X_train_Mou])
train_y = np.concatenate([train_y, y_train_Mou])


Berry_X, Berry_y =  create_input_and_output_data(df = df_Berry)
Oat_X, Oat_y =  create_input_and_output_data(df = df_Oat)


param = {'learning_rate': 0.31553117247348733,
         'max_delta_step': 1.7726044219753656,
         'max_depth': 10,
         'min_child_weight': 1.3845040588450772,
         'num_rounds': 342.68325188584106,
         'reg_alpha': 0.531395259755843,
         'reg_lambda': 3.744980563764689,
         'weight': 0.26187490421514203}

num_round = param["num_rounds"]
param["tree_method"] = "gpu_hist"
param["sampling_method"] = "gradient_based"
param['objective'] = 'binary:logistic'


weights =  np.array([param["weight"] if binding == 0 else 1.0 for binding in np.array(train_y)])


del param["num_rounds"]
del param["weight"]

dtrain = xgb.DMatrix(np.array(train_X), weight = weights, label = np.array(train_y),
            feature_names= feature_names)
dtest = xgb.DMatrix(np.array(test_X), label = np.array(test_y),
                    feature_names= feature_names)

dtest_new = xgb.DMatrix(np.array(test_new_X), label = np.array(test_new_y), feature_names= feature_names)
dOat = xgb.DMatrix(np.array(Oat_X), label = np.array(Oat_y), feature_names= feature_names)

evallist = [(dOat, 'eval'), (dtrain, 'train')]

bst = xgb.train(param,  dtrain, int(num_round),evallist, verbose_eval=1)
y_test_pred = np.round(bst.predict(dtest))
acc_test = np.mean(y_test_pred == np.array(test_y))
roc_auc = roc_auc_score(np.array(test_y), bst.predict(dtest))

print("Accuracy on test set: %s, ROC-AUC score for test set: %s"  % (acc_test, roc_auc))


y_test_new_pred = np.round(bst.predict(dtest_new))
acc_test_new = np.mean(y_test_new_pred == np.array(test_new_y))
roc_auc_new = roc_auc_score(np.array(test_new_y), bst.predict(dtest_new))

print("All enzymes:")
print("Accuracy on test set: %s, ROC-AUC score for test set: %s"  % (acc_test_new, roc_auc_new))

dBerry = xgb.DMatrix(np.array(Berry_X), label = np.array(Berry_y), feature_names= feature_names)

y_test_Berry_pred = np.round(bst.predict(dBerry))
acc_test_Berry = np.mean(y_test_Berry_pred == np.array(Berry_y))
roc_auc_Berry = roc_auc_score(np.array(Berry_y), bst.predict(dBerry))
print("Accuracy on Berry validation set: %s, ROC-AUC score for test set: %s"  % (acc_test_Berry, roc_auc_Berry))

dOat = xgb.DMatrix(np.array(Oat_X), label = np.array(Oat_y), feature_names= feature_names)

y_test_Oat_pred = np.round(bst.predict(dOat))
acc_test_Oat = np.mean(y_test_Oat_pred == np.array(Oat_y))
roc_auc_Oat = roc_auc_score(np.array(Oat_y), bst.predict(dOat))
print("Accuracy on Oat validation set: %s, ROC-AUC score for test set: %s"  % (acc_test_Oat, roc_auc_Oat))

from sklearn.metrics import matthews_corrcoef
matthews_corrcoef(Oat_y, np.round(bst.predict(dOat))), matthews_corrcoef(Berry_y, np.round(bst.predict(dBerry)))

### Adding no Yang et al. data to the training set:


df_train = pd.read_pickle(join(CURRENT_DIR, ".." ,"data",
                               "splits", "df_train_with_ESM1b_ts.pkl"))
df_train = df_train.loc[df_train["ESM1b_ts"] != ""]
df_train.reset_index(inplace = True, drop = True)


train_X, train_y =  create_input_and_output_data(df = df_train)
test_X, test_y =  create_input_and_output_data(df = df_test)

df_test_new = df_Mou.copy()
test_new_X, test_new_y =  create_input_and_output_data(df = df_test_new)



X_train_Mou = test_new_X
y_train_Mou = test_new_y


Berry_X, Berry_y =  create_input_and_output_data(df = df_Berry)
Oat_X, Oat_y =  create_input_and_output_data(df = df_Oat)


param = {'learning_rate': 0.31553117247348733,
         'max_delta_step': 1.7726044219753656,
         'max_depth': 10,
         'min_child_weight': 1.3845040588450772,
         'num_rounds': 342.68325188584106,
         'reg_alpha': 0.531395259755843,
         'reg_lambda': 3.744980563764689,
         'weight': 0.26187490421514203}

num_round = param["num_rounds"]
param["tree_method"] = "gpu_hist"
param["sampling_method"] = "gradient_based"
param['objective'] = 'binary:logistic'


weights =  np.array([param["weight"] if binding == 0 else 1.0 for binding in np.array(train_y)])


del param["num_rounds"]
del param["weight"]

dtrain = xgb.DMatrix(np.array(train_X), weight = weights, label = np.array(train_y),
            feature_names= feature_names)
dtest = xgb.DMatrix(np.array(test_X), label = np.array(test_y),
                    feature_names= feature_names)

dtest_new = xgb.DMatrix(np.array(test_new_X), label = np.array(test_new_y), feature_names= feature_names)
dOat = xgb.DMatrix(np.array(Oat_X), label = np.array(Oat_y), feature_names= feature_names)

evallist = [(dOat, 'eval'), (dtrain, 'train')]

bst = xgb.train(param,  dtrain, int(num_round),evallist, verbose_eval=1)
y_test_pred = np.round(bst.predict(dtest))
acc_test = np.mean(y_test_pred == np.array(test_y))
roc_auc = roc_auc_score(np.array(test_y), bst.predict(dtest))

print("Accuracy on test set: %s, ROC-AUC score for test set: %s"  % (acc_test, roc_auc))


y_test_new_pred = np.round(bst.predict(dtest_new))
acc_test_new = np.mean(y_test_new_pred == np.array(test_new_y))
roc_auc_new = roc_auc_score(np.array(test_new_y), bst.predict(dtest_new))

print("All enzymes:")
print("Accuracy on test set: %s, ROC-AUC score for test set: %s"  % (acc_test_new, roc_auc_new))

dBerry = xgb.DMatrix(np.array(Berry_X), label = np.array(Berry_y), feature_names= feature_names)

y_test_Berry_pred = np.round(bst.predict(dBerry))
acc_test_Berry = np.mean(y_test_Berry_pred == np.array(Berry_y))
roc_auc_Berry = roc_auc_score(np.array(Berry_y), bst.predict(dBerry))
print("Accuracy on Berry validation set: %s, ROC-AUC score for test set: %s"  % (acc_test_Berry, roc_auc_Berry))

dOat = xgb.DMatrix(np.array(Oat_X), label = np.array(Oat_y), feature_names= feature_names)

y_test_Oat_pred = np.round(bst.predict(dOat))
acc_test_Oat = np.mean(y_test_Oat_pred == np.array(Oat_y))
roc_auc_Oat = roc_auc_score(np.array(Oat_y), bst.predict(dOat))
print("Accuracy on Oat validation set: %s, ROC-AUC score for test set: %s"  % (acc_test_Oat, roc_auc_Oat))

from sklearn.metrics import matthews_corrcoef
matthews_corrcoef(Oat_y, np.round(bst.predict(dOat))), matthews_corrcoef(Berry_y, np.round(bst.predict(dBerry)))
