==> notebooks_and_code/4_1 - Training Graph Neural Network.ipynb <==
import sys
import os
from os.path import join
import numpy as np
import pickle
import random
import pandas as pd
import torch
import torch.nn as nn
from torch.utils.data import DataLoader
from torch.utils.data import Dataset
device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
print(device)

sys.path.append('.\\additional_code')
from xgboost_training import *

CURRENT_DIR = os.getcwd()
print(CURRENT_DIR)

## 1. Loading and preprocessing data:

df_train = pd.read_pickle(join(CURRENT_DIR, ".." ,"data","splits", "df_train_with_ESM1b_ts.pkl"))
df_test = pd.read_pickle(join(CURRENT_DIR, ".." ,"data", "splits", "df_test_with_ESM1b_ts.pkl"))

### (a) Balancing datasets such that we have same amount of negative as positive data samples:

pos_UIDs = df_train["Uniprot ID"].loc[df_train["Binding"] == 1]
for UID in pos_UIDs:
    n_pos = len(df_train.loc[df_train["Binding"] == 1].loc[df_train["Uniprot ID"] == UID])
    help_df = df_train.loc[df_train["Binding"] == 0].loc[df_train["Uniprot ID"] == UID]
    df_train.drop(list(help_df.index)[n_pos:], inplace= True)
    df_train.reset_index(inplace = True, drop = True)

pos_UIDs = df_test["Uniprot ID"].loc[df_test["Binding"] == 1]
for UID in pos_UIDs:
    n_pos = len(df_test.loc[df_test["Binding"] == 1].loc[df_test["Uniprot ID"] == UID])
    help_df = df_test.loc[df_test["Binding"] == 0].loc[df_test["Uniprot ID"] == UID]
    df_test.drop(list(help_df.index)[n_pos:], inplace= True)
    df_test.reset_index(inplace = True, drop = True)

### (b) Create dictionary with all target values

mol_files = list(set(df_train["molecule ID"])) + list(set(df_test["molecule ID"]))
mol_files = list(set(mol_files))

target_variable_dict = {}
target_variable_dict = create_target_dict(df = df_train, target_variable_dict = target_variable_dict)
target_variable_dict = create_target_dict(df = df_test, target_variable_dict = target_variable_dict)

### (c) Get list with input combinations of Uniprot ID and metabolite ID

train_IDs = get_uid_cid_IDs(df_train)
test_IDs = get_uid_cid_IDs(df_test)

print(np.mean(df_train["Binding"]), np.mean(df_test["Binding"]))
print(len(train_IDs), len(test_IDs))

## 2. Calculating input matrices for metabolites

### (a) Creating input matrices:

def calculate_and_save_input_matrixes(molecule_ID, save_folder = join(CURRENT_DIR, ".." ,"data", "substrate_data",
                                                                      "GNN_input_matrices")):
    molecule_ID = molecule_ID.replace(":", "_")
    molecule_ID = molecule_ID.replace("/", "Q")
    [XE, X, A] = create_input_data_for_GNN_for_substrates(substrate_ID = molecule_ID, print_error=True)
    if not A is None:
        np.save(join(save_folder, molecule_ID + '_X.npy'), X) #feature matrix of atoms/nodes
        np.save(join(save_folder, molecule_ID + '_XE.npy'), XE) #feature matrix of atoms/nodes and bonds/edges
        np.save(join(save_folder, molecule_ID + '_A.npy'), A) #adjacency matrix
        
        
def calculate_atom_and_bond_feature_vectors(mol_files):
    #check if feature vectors have already been calculated:
    try:
        os.mkdir(join(CURRENT_DIR, ".." ,"data", "substrate_data", "mol_feature_vectors"))
    except FileExistsError:
        None
    
    #existing feature vector files:
    feature_files = os.listdir(join(CURRENT_DIR, ".." ,"data", "substrate_data", "mol_feature_vectors"))
    for mol_file in mol_files:
        #check if feature vectors were already calculated:
        try:
            if not mol_file + "-atoms.txt" in  feature_files:
                #load mol_file
                is_CHEBI_ID = (mol_file[0:5] == "CHEBI")
                is_inchi = (mol_file[0:5] == "InChI")
                if is_CHEBI_ID:
                    ID = int(mol_file.split(" ")[0].split(":")[-1])
                    Inchi = list(df_chebi_to_inchi["Inchi"].loc[df_chebi_to_inchi["ChEBI"] == float(ID)])[0]

                    if not pd.isnull(Inchi):
                        mol = Chem.inchi.MolFromInchi(Inchi)
                    else:
                        print(ID, Inchi)
                elif is_inchi:
                    mol = Chem.inchi.MolFromInchi(mol_file)
                    mol_file = mol_file.replace("/", "Q")
                else:
                    mol = Chem.MolFromMolFile(mol_folder +  "/mol-files/" + mol_file + '.mol')
       
            if not mol is None:
                calculate_atom_feature_vector_for_mol_file(mol, mol_file)
                calculate_bond_feature_vector_for_mol_file(mol, mol_file)
        except OSError: pass

calculate_atom_and_bond_feature_vectors(mol_files = mol_files)

for mol_ID in mol_files:
    calculate_and_save_input_matrixes(molecule_ID = mol_ID)

###  (b) Removing all datapoints without molecule input file:

df_train["molecule ID"] = [ID.replace(":", "_") for ID in df_train["substrate ID"]]
df_test["molecule ID"] = [ID.replace(":", "_") for ID in df_test["substrate ID"]]

valid_mols = os.listdir(join(CURRENT_DIR, ".." ,"data", "substrate_data", "GNN_input_matrices"))
valid_mols = [mol.split("_A")[0] for mol in valid_mols]

df_train = df_train.loc[df_train["molecule ID"].isin(valid_mols)]
df_test = df_test.loc[df_test["molecule ID"].isin(valid_mols)]

train_IDs = get_uid_cid_IDs(df_train)
test_IDs = get_uid_cid_IDs(df_test)
df_train

### (c) Creating representations for the enzymes:

uids_list = list(set(df_train["Uniprot ID"])) + list(set(df_test["Uniprot ID"]))
uids_list = list(set(uids_list))
uid_to_emb = {}
embeddings = np.zeros((0,1280))
for uid in uids_list:
    try:
        emb = np.reshape(np.array(list(df_train["ESM1b"].loc[df_train["Uniprot ID"] == uid])[0]), (1,1280))
    except IndexError:
        try:
            emb = np.reshape(np.array(list(df_test["ESM1b"].loc[df_test["Uniprot ID"] == uid])[0]), (1,1280))
        except IndexError:
            emb = np.reshape(np.array(list(df_validation["ESM1b"].loc[df_validation["Uniprot ID"] == uid])[0]), (1,1280))
    embeddings = np.concatenate([embeddings, emb])
    uid_to_emb[uid] = emb

We perform a PCA an the enzyme representations to get 50-dimensional representations

from sklearn.decomposition import PCA
dim = 50

pca = PCA(n_components = dim)
pca.fit(embeddings)
emb_pca = pca.transform(embeddings)

#Calculate mean and std to normalize the PCA-transformed vectors
mean = np.mean(emb_pca, axis = 0)
std = np.std(emb_pca, axis = 0)

uid_to_pca_emb = {}

for i, uid in enumerate(uids_list):
    uid_to_pca_emb[uid] = (emb_pca[i] - mean) / std

uid_to_emb = uid_to_pca_emb

## 3. Training GNN:

###  (a) Defining a DataGenerator:

class CustomDataSet(Dataset):
    def __init__(self, split_IDs, folder):
        self.all_IDs = split_IDs
        self.folder = folder

    def __len__(self):
        return len(self.all_IDs)

    def __getitem__(self, idx):
        ID = self.all_IDs[idx]
        try:
            [uid,cid1, cid2] = ID.split("_") 
            cid = cid1 +"_"+cid2
        except ValueError:
            [uid,cid] = ID.split("_")
            
        XE = torch.tensor(np.load(join(self.folder, cid + '_XE.npy')), dtype = torch.float32)
        X = torch.tensor(np.load(join(self.folder, cid + '_X.npy')), dtype = torch.float32)
        A = torch.tensor(np.load(join(self.folder, cid + '_A.npy')), dtype = torch.float32)
        ESM1b = torch.tensor(uid_to_emb[uid], dtype = torch.float32)
        label = torch.tensor(target_variable_dict[ID], dtype= torch.float32)
        return XE,X,A,ESM1b, label

### (b) Splitting the training set in a validation and a training set:

n = len(train_IDs) 
random.seed(1)
random.shuffle(train_IDs)
test_IDs = train_IDs[int(0.8*n):]
train_IDs = train_IDs[:int(0.8*n)]


batch_size = 64

train_dataset = CustomDataSet(folder  = join(CURRENT_DIR, ".." ,"data", "substrate_data",
                                             "GNN_input_matrices"), split_IDs = train_IDs)
train_loader = DataLoader(train_dataset , batch_size=batch_size, shuffle=True, drop_last=True)

test_dataset = CustomDataSet(folder  = join(CURRENT_DIR, ".." ,"data", "substrate_data",
                                            "GNN_input_matrices"), split_IDs = test_IDs)
test_loader = DataLoader(test_dataset , batch_size=batch_size, shuffle=False, drop_last=True)

n_train_batches = int(len(train_dataset)/batch_size)
n_test_batches = int(len(test_dataset)/batch_size)
train_batches = list(range(n_train_batches))
test_batches = list(range(n_test_batches))

### (c) Training GNN:

pre_training = True

import torch.optim as optim

model = GNN(D= 100, N = 70, F1 = 32 , F2 = 10, F = F1+F2).to(device)
if pre_training:
    model.load_state_dict(torch.load(join(CURRENT_DIR, ".." ,"data", "substrate_data_KM", "GNN", "Pytorch_GNN_KM")))

criterion = nn.BCELoss()
optimizer = optim.Adam(model.parameters(), lr=0.001, weight_decay= 0.00001)


    


for epoch in range(10):  # loop over the dataset multiple times
    model.train()
    running_loss = 0.0
    for i, [XE, X, A,ESM1b, labels] in enumerate(train_loader):
        # zero the parameter gradients
        optimizer.zero_grad()
        XE, X, A, ESM1b, labels = XE.to(device), X.to(device), A.to(device),ESM1b.to(device), labels.to(device)
        # forward + backward + optimize
        outputs = model(XE, X, A, ESM1b)
        loss = criterion(outputs, labels.view((batch_size,-1)))
        loss.backward()
        optimizer.step()

        # print statistics
        running_loss += loss.item()
        if i % 20 == 19:    # print every 2000 mini-batches
            print('[%d, %5d] loss: %.3f' %
                  (epoch + 1, i + 1, running_loss / 20))
            running_loss = 0.0
            
    #After each epoch, calculate the validation loss:
    running_acc = 0.0
    running_loss = 0.0
    model.eval()
    for i, [XE, X, A,ESM1b, labels] in enumerate(test_loader):
        XE, X, A, ESM1b, labels = XE.to(device), X.to(device), A.to(device),ESM1b.to(device), labels.to(device)
        
        with torch.no_grad():
            outputs = model(XE, X, A, ESM1b)
        loss = criterion(outputs, labels.view((batch_size,-1)))
        running_loss += loss.item()

        outputs2 = np.round(outputs.view(-1).cpu().detach().numpy()) 
        labels2 = labels.cpu().detach().numpy()
        acc = np.mean([outputs2[i] == labels2[i] for i in range(len(labels))])
        running_acc += acc

    print("Epoch: %s, Val. loss: %s, Val. acc: %s" % (epoch, np.round(running_loss/(i+1),2),
                                                                  np.round(running_acc/(i+1), 2)))

print('Finished Training')

torch.save(model.state_dict(),join(CURRENT_DIR, ".." ,"data", "substrate_data", "GNN", "Pytorch_GNN_with_pretraining"))

### 4. Creating GNN representations for traing and test set:

Load trained model:

model = GNN(D= 100, N = 70, F1 = 32 , F2 = 10, F = F1+F2).to(device)
model.load_state_dict(torch.load(join(CURRENT_DIR, ".." ,"data", "substrate_data", "GNN", "Pytorch_GNN_with_pretraining")))
model.eval()

Loading data:

df_train = pd.read_pickle(join(CURRENT_DIR, ".." ,"data","splits", "df_train_with_ESM1b_ts.pkl"))
df_test = pd.read_pickle(join(CURRENT_DIR, ".." ,"data", "splits", "df_test_with_ESM1b_ts.pkl"))

Creating a DataFrame with all metabolites in our dataset:

valid_mols = os.listdir(join(CURRENT_DIR, ".." ,"data", "substrate_data", "GNN_input_matrices"))
valid_mols = [mol.split("_A.npy")[0] for mol in valid_mols]
valid_mols = [mol for mol in valid_mols if not "." in mol]
df_mols = pd.DataFrame(data = {"molecule ID" : valid_mols})

#To create the substrate rep, the UID does not matter. Therfore, setting it random:
df_mols["uid"] = "P9WIQ3"
df_mols

input_data_folder = join(CURRENT_DIR, ".." ,"data", "substrate_data", "GNN_input_matrices")

def get_representation_input(cid_list):
    XE = ();
    X = ();
    A = ();
    UniRep = ();
    extras = ();
    # Generate data
    for i in range(len(cid_list)):
        cid  = cid_list[i]

        X = X + (np.load(join(input_data_folder, cid + '_X.npy')), );
        XE = XE + (np.load(join(input_data_folder, cid + '_XE.npy')), );
        A = A + (np.load(join(input_data_folder, cid + '_A.npy')), );
    return(XE, X, A)


def get_substrate_representations(df):
    df["substrate_rep"] = ""
    
    i = 0
    n = len(df)

    while i*64 <= n:
        cid_all = list(df["molecule ID"])

        if (i+1)*64  <= n:
            XE, X, A= get_representation_input(cid_all[i*64:(i+1)*64])
            
            XE = torch.tensor(np.array(XE), dtype = torch.float32).to(device)
            X = torch.tensor(np.array(X), dtype = torch.float32).to(device)
            A = torch.tensor(np.array(A), dtype = torch.float32).to(device)
            representations = model.get_GNN_rep(XE, X,A).cpu().detach().numpy()
            df["substrate_rep"][i*64:(i+1)*64] = list(representations[:, :D])
        else:
            print(i)
            XE, X, A= get_representation_input(cid_all[i*64:(i+1)*64])
            XE = torch.tensor(np.array(XE), dtype = torch.float32).to(device)
            X = torch.tensor(np.array(X), dtype = torch.float32).to(device)
            A = torch.tensor(np.array(A), dtype = torch.float32).to(device)
            representations = model.get_GNN_rep(XE, X,A).cpu().detach().numpy()
            df["substrate_rep"][-len(representations):] = list(representations[:, :D])
        i += 1
    return(df)

df_mols = get_substrate_representations(df = df_mols)
df_mols

df_train = pd.read_pickle(join(CURRENT_DIR, ".." ,"data", "splits", "df_train_with_ESM1b_ts_GNN.pkl"))
df_test = pd.read_pickle(join(CURRENT_DIR, ".." ,"data", "splits", "df_test_with_ESM1b_ts_GNN.pkl"))

df_train["GNN rep (pretrained)"] = ""
for ind in df_train.index:
    try:
        df_train["GNN rep (pretrained)"][ind] = list(df_mols["substrate_rep"].loc[df_mols["molecule ID"] == df_train["molecule ID"][ind].replace(":", "_").replace("Q", "/")])[0]
    except IndexError:
        pass

df_test["GNN rep (pretrained)"] = ""
for ind in df_test.index:
    try:
        df_test["GNN rep (pretrained)"][ind] = list(df_mols["substrate_rep"].loc[df_mols["molecule ID"] == df_test["molecule ID"][ind].replace(":", "_").replace("Q", "/")])[0]
    except IndexError:
        pass

df_train.to_pickle(join(CURRENT_DIR, ".." ,"data", "splits", "df_train_with_ESM1b_ts_GNN.pkl"), protocol = 4)
df_test.to_pickle(join(CURRENT_DIR, ".." ,"data", "splits", "df_test_with_ESM1b_ts_GNN.pkl"), protocol = 4)

#df_engqvist.to_pickle(join(CURRENT_DIR, "alex_data", "new_test_data_Engqvist_group_with_GNN_train_test_split_with_GNN_similar.pkl"), protocol = 4)
